Minor tweaks to hex layout

+++ ui/hexmap/GUIEWHex.java
@@ -13,7 +13,7 @@ import game.*;
 
 /**
  * Class GUIBattleHex holds GUI info for one hex with E-W orientation.
- * @version $Id: GUIEWHex.java,v 1.1 2005/07/20 12:44:51 wakko666 Exp $
+ * @version $Id: GUIEWHex.java,v 1.2 2005/07/21 11:15:23 wakko666 Exp $
  * @author David Ripton
  * @author Romain Dolbeau
  */
@@ -35,8 +35,8 @@ public class GUIEWHex extends GUIHex
     // A1-A3, B1-B4, C1-C5, D1-D6, E1-E5, F1-F4.
     // Letters increase left to right; numbers increase bottom to top.
 
-    public GUIEWHex(int cx, int cy, int scale, Component map,
-        int xCoord, int yCoord)
+    public GUIEWHex(double cx, double cy, int scale, Component map,
+        double xCoord, double yCoord)
     {
         super(new BattleHex(xCoord, yCoord));
         this.map = map;
